+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Fix escape handling in Safari
+
 ## 1.3.0 - 2024 Jan 10
 - Add `class` property
 - Fix escape shortcut handling